requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
chardet==4.0.0
cleo==0.8.1
clikit==0.6.2
colorama==0.4.4
colormap==1.0.3
crashtest==0.3.1
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
html5lib==1.1
idna==2.10
imageio==2.9.0
iniconfig==1.1.1
keyring==21.8.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.4.2
msgpack==1.0.2
numpy==1.20.3
packaging==20.9
pastel==0.2.1
pathlib==1.0.1
pexpect==4.8.0
Pillow==8.2.0
pipenv==2020.11.15
pkginfo==1.7.0
pluggy==0.13.1
poetry==1.1.6
poetry-core==1.0.3
ptyprocess==0.7.0
py==1.10.0
pylev==1.3.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
requests==2.25.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
toml==0.10.2
tomlkit==0.7.2
urllib3==1.26.4
virtualenv==20.4.6
virtualenv-clone==0.5.4
webencodings==0.5.1
XlsxWriter==1.4.3
